renamed:    src/lexer/liLexer.x -> src/Lexer/Lexer.x new file:   src/Main.hs new file:   src/Parser/SParser.hs deleted:    src/lexer/sampleIO/out.alex deleted:    src/lexer/sampleIO/test.scm deleted:    src/parser/sampleParser.hs
Modifying the Package structure to support the hierarchien build and
increase the modularity of the project

Adding the file SParser.hs which includes basic parsers for the Tok
data structure defines in Lexer.xs

Also modified the file Lexer.x to support the Package structure

Adding Main.hs file which can be used to test different parsers and
other functions defined in other modules instead to touching the modules
itself.

Now in order to build Just do 'ghc --make Main.hs' in the "li/src/"
directory

TODO :: Currently, when one does a 'ghc --make Main.hs' in the "li/src/"
it pollutes the directories by scattering the binaries in the src
itlself. It has to be fixed.